requirements.txt
-e git://github.com/django/django.git@6a4f13de27e2c46b6e516a01ba77fd4fe03d9dbe#egg=django
Pillow==3.1.0
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
gunicorn==19.4.5
psycopg2==2.6.1
static3==0.6.1
whitenoise==2.0.6
mistune==0.2.0
